+++ internal/models/price_set.go
@@ -5,16 +5,16 @@ type PriceSet struct {
 	Name            string    `json:"name"`
 	CategoryID      int       `json:"category_id"`
 	SubcategoryID   int       `json:"subcategory_id"`
-	Quantity        int       `json:"quantity"`
+	Quantity        float64   `json:"quantity"`
 	Price           int       `json:"price"`
 	SubcategoryName string    `json:"subcategory_name,omitempty"`
 	Items           []SetItem `json:"items"`
 }
 
 type SetItem struct {
-	ID         int    `json:"id"`
-	PriceSetID int    `json:"price_set_id"`
-	ItemID     int    `json:"item_id"`
-	Quantity   int    `json:"quantity"`
-	ItemName   string `json:"item_name,omitempty"`
+	ID         int     `json:"id"`
+	PriceSetID int     `json:"price_set_id"`
+	ItemID     int     `json:"item_id"`
+	Quantity   float64 `json:"quantity"`
+	ItemName   string  `json:"item_name,omitempty"`
 }
